Interface manutenção Cotações

+++ app/src/main/java/com/jacksonasantos/travelplan/dao/interfaces/CurrencyQuoteIDAO.java
@@ -6,10 +6,10 @@ import java.util.Date;
 import java.util.List;
 
 public interface CurrencyQuoteIDAO {
-    CurrencyQuote fetchCurrencyQuoteById(int currencyQuoteId);
+    CurrencyQuote fetchCurrencyQuoteById(Long currencyQuoteId);
     List<CurrencyQuote> fetchAllCurrencyQuotes();
     boolean addCurrencyQuote(CurrencyQuote currencyQuote);
-    boolean deleteCurrencyQuote(int currencyQuoteId);
+    boolean deleteCurrencyQuote(Long currencyQuoteId);
     boolean updateCurrencyQuote(CurrencyQuote currencyQuote);
 
     CurrencyQuote findQuoteDay(int currency_type, Date quote_date);
